refactor: move contracts to this repo temporaly

+++ src/main/java/com/edu/upc/ilanguagesubscription/command/application/dto/RegisterSubscriptionRequestDto.java
@@ -3,14 +3,14 @@ package com.edu.upc.ilanguagesubscription.command.application.dto;
 public class RegisterSubscriptionRequestDto {
     private int subscriptionId;
     private String name;
-    private float price;
+    private int price;
     private int monthDuration;
 
     public String getName() {
         return name;
     }
 
-    public float getPrice() {
+    public int getPrice() {
         return price;
     }
