+++ src/main/java/com/edu/upc/ilanguagesubscription/command/domain/contracts/events/SubscriptionRegistered.java
@@ -9,6 +9,6 @@ public class SubscriptionRegistered {
     int subscriptionId;
     String name;
     int monthDuration;
-    float price;
+    int price;
     Instant occurredOn;
 }
